fix: Moving middleware dependencies to middleware folder
