+++ app/interfaces/api/v1/controllers/sync_controller.py
@@ -7,8 +7,8 @@ from typing import Annotated
 from fastapi import APIRouter, Depends
 
 from app.dependencies.celery_dependencies import get_celery
-from app.dependencies.middleware_dependencies import get_current_user_id
 from app.exceptions.application_exception import ApplicationError
+from app.middleware.middleware_dependencies import get_current_user_id
 from app.schemas.sync_schema import (
     SyncBulkRequest,
     SyncBulkResponse,
