+++ app/middleware/user_admin_middleware.py
@@ -8,9 +8,9 @@ from typing import Annotated
 
 from fastapi import Depends
 
-from app.dependencies.middleware_dependencies import get_current_user_id
 from app.dependencies.services_dependencies import get_group_service
 from app.exceptions.application_exception import ApplicationError
+from app.middleware.middleware_dependencies import get_current_user_id
 from app.services.group_service import GroupService
 from app.utils.create_exception_util import create_http_exception
